Fix a corner case with scheduling queue-processing GSource
There was an assumption that when SoupMessageQueueItem's async_pending
flag is set to TRUE, there is no need to schedule the idle GSource for
the GMainContext the item has.

Apparently, in some corner case, this assumption is wrong. It is
possible that the async_pending flag is set to TRUE, but there is no
idle GSource scheduled to process the item. This may not often be a
problem, since there could be other items with the same GMainContext
earlier in the queue, so the GSource would be scheduled.

The corner case I had is when I tried to do "flatpak install". Flatpak
downloaded the delta superblock file and then scheduled more files to
be downloaded, but nothing happened. That's because the
SoupMessageQueueItem for the superblock never left the queue (even if
the download was finished) - it remained there as a first item in the
queue with the async_pending flag set to TRUE and with the state
SOUP_MESSAGE_FINISHING. So the soup_session_real_kick_queue assumed
that the context already had an idle GSource scheduled and skipped
that step for the rest of the items with the same GMainContext in the
queue.

The proper fix would be investigating and fixing why this inconsistent
state ever happened, but it was hard for me to debug (it's slow, so
the connections get interrupted, terminated and whatever). Also, the
code is a bit of a maze, so it does not help either.

So the workaround is to avoid checking the async_pending flag as a way
of ensuring that the GSource was scheduled, but actually to check if
the GSource is still there.

I'm afraid this is more or less a rewrite of the code that was there
before the commit 8e1b0f5a99b1f1236c3d1124cc718e905e5741f1.